feat: making desktop login less confusing
This is a total hack but trying to make the desktop login process a
little less confusing. When we first load the UI from twitch giving us a
token, we attempt to put our token on the clipboard. Then we observe the
clipboard for the token disappearing. We assume the clipboard being
cleared means the application logged in successfully.

closes nullinside-development-group/twitch-streaming-tools#56